Fix main menu display on a secondary monitor
When Taiga is placed on a secondary display that is at the left side of the main one, using LOWORD & HIWORD macros to extract the screen coordinates didn't work, probably because they were supposed to be negative numbers. I'm not sure if this resolves the issue; I don't have a second monitor to test.