Fix string encoder
The previous code did not take into account that the size of the data can change after converting it to std::string. This resulted in filters such as "[Fansub] Tokyo Ghoul √A" to be encoded to invalid filter strings.

Since the new struct is incompatible with the previous one, version is incremented to 0x01.