Allow changing status to Completed after manually changing episode number
The previous change in 2ae15694ff29173bce9e3605bf61a9bb08770c70 was useful when people updated the episode number of an on-hold series and it was no longer moved to Currently Watching. However, most of the time you'd expect the series to be set as Completed when you watch the last episode. So we're finding a middle ground here with this change.