Use HTTPS with MyAnimeList
MAL has recently added HTTPS/SSL support to their website. By mid-September, their API is going to be HTTPS-only. Prior to the public announcement, I was told that they won't be redirecting HTTP requests to the API with a 301 response, for security reasons. In other words, we have to release an update by then.

Closes #295.