Do not restart Settings dialog after restoring defaults
EndDialog does not destroy the dialog box immediately, so we end up with two Settings dialogs, where the old one is hidden. This causes issues such as certain pages not being displayed on the new dialog.

We could probably work around this with a custom PostMessage to the main window. But restoring default settings is not a common action, so we will avoid this complexity for now.