Set MACHINE option to x86 while building libcurl
The default value for `MACHINE` is supposed to be `x86` according to curl's BUILD.WINDOWS.txt, and I haven't had any issues building it on a 64-bit system so far. But apparently it doesn't work in every case due to how %PROCESSOR_ARCHITECTURE% environment variable behaves.

Reported by @tophf at #294.