Check paint struct before painting windows
Discord's screen sharing feature somehow causes Taiga to receive WM_PAINT messages with an empty update region, which results in a null pointer for the paint struct.

I was able to reproduce the issue on PageBaseInfo::OnPaint specifically, but any OnPaint function using the pointer should check it first to be safe.

Fixes #696.