Update Jellyfin Integration
The current Version of the Jellyfin integration had an erroneous ! in the Regex, which has been removed in this commit.
Also allow finding it, in the same way as Plex is detected, via IP&Port (Portrange between 1-99999, as Jellyfin allows you to set the port on which it is running).